requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=5.4.1=pypi_0
attrs=24.2.0=pypi_0
blinker=1.8.2=pypi_0
ca-certificates=2024.9.24=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
narwhals=1.13.2=pypi_0
numpy=2.0.2=pypi_0
openssl=3.0.15=h827c3e9_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py39haa95532_0
protobuf=5.28.3=pypi_0
pyarrow=18.0.0=pypi_0
pycryptodome=3.21.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
python=3.9.20=h8205438_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.21.0=pypi_0
setuptools=75.1.0=py39haa95532_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
streamlit=1.40.0=pypi_0
tenacity=9.0.0=pypi_0
toml=0.10.2=pypi_0
tornado=6.4.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
watchdog=5.0.3=pypi_0
wheel=0.44.0=py39haa95532_0
